Fixed the bug, now only error with big ints

+++ PayRoll/Employee.java
@@ -3,10 +3,10 @@ package PayRoll;
 public class Employee extends Person{
 private int PayGrade;
 private double Salary;
-private int StockAmount;
+private double StockAmount;
 private boolean PayedInShares;
 
-public Employee(int PayG, double Sal, int Stock, String Called, boolean Shares)
+public Employee(int PayG, double Sal, double Stock, String Called, boolean Shares)
 {
 	PayGrade = PayG;
 	Salary = Sal;
@@ -22,7 +22,7 @@ public String getEmployeeName()
 {
 	return getName();
 }
-public int getStock()
+public double getStock()
 {
 	return StockAmount;
 }
